refactor #55 : 전체적인 구조 변경
