+++ src/Pages/MinePage/AojiLog.jsx
@@ -1,54 +1,32 @@
-import { differenceInSeconds, parseISO } from 'date-fns';
+import React from 'react';
+import { NoData } from 'Components';
 
-import { format } from 'date-fns';
+import AojiLogData from './AojiLogData';
 
-const fotmatDate = date => {
-  if (!date) return '공부 중!';
-  date = new Date(date);
-
-  return format(date, 'HH:mm:ss');
-};
-
-const fotmatRecord = record => {
-  if (record === 'doing study') return '00:00:00';
-  const arr = record.split(':');
-  const newRecord = arr.map(i => {
-    return parseInt(i) < 10 ? '0' + i : i;
-  });
-
-  return newRecord.join(':');
-};
-const AojiLog = ({ data }) => {
-  const earnedPoint = (
-    differenceInSeconds(parseISO(data.endAt), parseISO(data.startAt)) /
-    60 /
-    60 /
-    8.0
-  ).toFixed(2);
-
-  const validClick = isEnd => {
-    if (isEnd)
-      return (
-        <div className="button valid" onClick={handleEditLog}>
-          수정
-        </div>
-      );
-    else return <div className="button not-valid">수정</div>;
-  };
-
-  const handleEditLog = () => {
-    alert('준비중입니다!');
-  };
+function AojiLog({ aojiLogs, attendScore }) {
   return (
-    <div className="row">
-      <div>{fotmatDate(data.startAt)}</div>
-      <div>{fotmatDate(data.endAt)}</div>
-      <div>{fotmatRecord(data.recodeTime)}</div>
-      <div>{!isNaN(earnedPoint) ? <>{earnedPoint}점</> : <>공부 중!</>}</div>
+    <div className="log box">
+      <div className="header">⛏️ 보충학습 기록</div>
+      <div className="temp row">
+        <div>시작 시간</div>
+        <div>종료 시간</div>
+        <div>공부 시간</div>
+        <div>차감 점수</div>
+        <div>수정</div>
+      </div>
 
-      <div> {validClick(data.endAt)}</div>
+      <div className="body">
+        {aojiLogs && aojiLogs.length ? (
+          aojiLogs.map(log => {
+            return <AojiLogData data={log} key={log.aojiTimeIndex} />;
+          })
+        ) : (
+          <NoData code={1} />
+        )}
+      </div>
+      <div className="score">현재 출결 점수 : {attendScore}</div>
     </div>
   );
-};
+}
 
 export default AojiLog;
